+++ packages/eslint-plugin/configs/i18n.js
@@ -1,0 +1,12 @@
+module.exports = {
+	plugins: ['@10up'],
+	rules: {
+		'@10up/valid-sprintf': 'error',
+		'@10up/i18n-translator-comments': 'error',
+		'@10up/i18n-text-domain': 'error',
+		'@10up/i18n-no-collapsible-whitespace': 'error',
+		'@10up/i18n-no-placeholders-only': 'error',
+		'@10up/i18n-no-variables': 'error',
+		'@10up/i18n-ellipsis': 'error',
+	},
+};